Move registration initialization functionality out of routes and into root file
